robonix/brain/requirements.txt
sensor-msgs-py==4.9.0
rosidl-cli==3.1.6
teleop-twist-keyboard==2.4.0
osrf-pycommon==2.1.6
rqt-console==2.0.3
rqt-action==2.0.1
rqt-topic==1.5.0
rqt-publisher==1.5.0
ament-index-python==1.4.0
rqt-graph==1.3.1
rqt-msg==1.2.0
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-bag==1.1.5
rqt-plot==1.1.5
rqt-bag-plugins==1.1.5
rqt-reconfigure==1.1.2
launch-xml==1.0.9
launch==1.0.9
launch-yaml==1.0.9
launch-testing==1.0.9
rqt-service-caller==1.0.5
rqt-srv==1.0.3
rqt-py-console==1.0.2
rqt-shell==1.0.2
tf2-ros-py==0.25.13
tf2-tools==0.25.13
action-tutorials-py==0.20.5
demo-nodes-py==0.20.5
quality-of-service-demo-py==0.20.5
topic-monitor==0.20.5
launch-ros==0.19.9
ros2launch==0.19.9
launch-testing-ros==0.19.9
ros2doctor==0.18.12
ros2service==0.18.12
ros2run==0.18.12
ros2interface==0.18.12
ros2lifecycle==0.18.12
ros2cli==0.18.12
ros2pkg==0.18.12
ros2component==0.18.12
ros2topic==0.18.12
ros2action==0.18.12
ros2param==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2bag==0.15.14
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
examples-rclpy-executors==0.15.3
ament-package==0.14.1
ament-lint==0.12.12
ament-cpplint==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
ament-lint-cmake==0.12.12
ament-cppcheck==0.12.12
ament-pep257==0.12.12
ament-copyright==0.12.12
ament-flake8==0.12.12
sros2==0.10.6
domain-coordinator==0.10.0
rosidl-runtime-py==0.9.3
rpyutils==0.2.1
rcutils==5.1.6
std-msgs==4.9.0
nav-msgs==4.9.0
visualization-msgs==4.9.0
geometry-msgs==4.9.0
sensor-msgs==4.9.0
diagnostic-msgs==4.9.0
trajectory-msgs==4.9.0
stereo-msgs==4.9.0
actionlib-msgs==4.9.0
shape-msgs==4.9.0
std-srvs==4.9.0
message-filters==4.3.7
rclpy==3.3.16
image-geometry==3.2.1
cv-bridge==3.2.1
rosidl-generator-c==3.1.6
rosidl-typesupport-introspection-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rosidl-parser==3.1.6
rosidl-adapter==3.1.6
rosidl-cmake==3.1.6
resource-retriever==3.1.3
laser-geometry==2.4.0
interactive-markers==2.3.2
qt-gui-py-common==2.2.4
qt-gui-cpp==2.2.4
qt-gui==2.2.4
qt-dotgraph==2.2.4
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-fastrtps-c==2.2.2
unique-identifier-msgs==2.2.1
map-msgs==2.1.0
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-c==2.0.2
angles==1.15.0
rmw-dds-common==1.6.0
turtlesim==1.4.2
ament-cmake-test==1.3.11
lifecycle-msgs==1.2.1
statistics-msgs==1.2.1
rosgraph-msgs==1.2.1
composition-interfaces==1.2.1
action-msgs==1.2.1
builtin-interfaces==1.2.1
rcl-interfaces==1.2.1
rqt-py-common==1.1.7
python-qt-binding==1.1.2
pcl-msgs==1.0.0
tf2-geometry-msgs==0.25.13
tf2-msgs==0.25.13
tf2-kdl==0.25.13
tf2-py==0.25.13
tf2-sensor-msgs==0.25.13
logging-demo==0.20.5
action-tutorials-interfaces==0.20.5
pendulum-msgs==0.20.5
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosidl-generator-py==0.14.4
example-interfaces==0.9.3
jsonschema-specifications==2025.4.1
pyzmq==27.0.0
attrs==25.3.0
packaging==25.0
rich==14.0.0
ipython==8.37.0
jupyter-client==8.6.3
psutil==7.0.0
ipykernel==6.29.5
tornado==6.5.1
traitlets==5.14.3
jupyter-core==5.8.1
decorator==5.2.1
tqdm==4.67.1
jsonschema==4.24.0
typing-extensions==4.14.0
anyio==4.9.0
pexpect==4.9.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
asttokens==3.0.0
markdown-it-py==3.0.0
pydantic-core==2.33.2
pygments==2.19.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
sse-starlette==2.3.6
executing==2.2.0
openai==1.93.2
mcp==1.10.1
debugpy==1.8.14
websocket-client==1.8.0
nest-asyncio==1.6.0
sniffio==1.3.1
exceptiongroup==1.3.0
python-dotenv==1.1.1
httpcore==1.0.9
anthropic==0.57.1
starlette==0.47.1
referencing==0.36.2
uvicorn==0.35.0
httpx==0.28.1
rpds-py==0.25.1
jedi==0.19.2
h11==0.16.0
jiter==0.10.0
parso==0.8.4
annotated-types==0.7.0
ptyprocess==0.7.0
stack-data==0.6.3
typing-inspection==0.4.1
httpx-sse==0.4.1
wcwidth==0.2.13
pure-eval==0.2.3
comm==0.2.2
matplotlib-inline==0.1.7
mdurl==0.1.2
python-multipart==0.0.20
ubuntu-pro-client==8001
pytz==2022.1
certifi==2020.6.20
systemd-python==234
setuptools==59.6.0
cloud-init==24.4
keyring==23.5.0
Twisted==22.1.0
pip==22.0.2
incremental==21.3.0
hyperlink==21.0.0
pyOpenSSL==21.0.0
Automat==20.2.0
service-identity==18.1.0
constantly==15.1.0
unicodedata2==14.0.0
PyQt5-sip==12.9.1
Pillow==9.0.1
more-itertools==8.10.0
click==8.0.3
pycurl==7.44.1
pytest==6.2.5
coverage==6.2
pydocstyle==6.1.1
PyQt5==5.15.6
PyYAML==5.4.1
zope.interface==5.4.0
configobj==5.0.6
fonttools==4.29.1
beautifulsoup4==4.10.0
lxml==4.8.0
importlib-metadata==4.6.4
flake8==4.0.1
chardet==4.0.0
PyGObject==3.42.1
ply==3.11
matplotlib==3.5.1
pyserial==3.5
cryptography==3.4.8
empy==3.3.4
SecretStorage==3.3.1
roman==3.3
idna==3.3
bcrypt==3.2.0
oauthlib==3.2.0
lz4==3.1.3+dfsg
mpi4py==3.1.3
Jinja2==3.0.3
pytest-cov==3.0.0
requests==2.25.1
Babel==2.8.0
pycodestyle==2.8.0
fs==2.4.12
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4
pyflakes==2.4.0
soupsieve==2.3.1
PyJWT==2.3.0
snowballstemmer==2.2.0
PyHamcrest==2.0.2
MarkupSafe==2.0.1
jsonpointer==2.0
jsonpatch==1.32
urllib3==1.26.5
numpy==1.21.5
pycairo==1.20.1
six==1.16.0
cov-core==1.15.0
launchpadlib==1.10.16
py==1.10.0
sympy==1.9
argcomplete==1.8.1
scipy==1.8.0
distro==1.7.0
rospkg-modules==1.6.0
appdirs==1.4.4
pydot==1.4.2
blinker==1.4
wadllib==1.3.6
kiwisolver==1.3.2
dbus-python==1.2.18
lark==1.1.1
iniconfig==1.1.1
distro-info==1.1+ubuntu0.2
html5lib==1.1
Brotli==1.0.9
lazr.uri==1.0.6
rosdistro-modules==1.0.1
zipp==1.0.0
catkin-pkg-modules==1.0.0
olefile==0.46
wheel==0.37.1
ufw==0.36.1
rosdep-modules==0.25.1
rosdep==0.25.1
httplib2==0.20.2
colcon-core==0.19.0
pyrsistent==0.18.1
docutils==0.17.1
lazr.restfulclient==0.14.4
ufoLib2==0.13.1
pluggy==0.13.0
cycler==0.11.0
netifaces==0.11.0
toml==0.10.2
pythran==0.10.0
nose2==0.9.2
jeepney==0.7.1
mccabe==0.6.1
gast==0.5.2
webencodings==0.5.1
colcon-zsh==0.5.0
colcon-bash==0.5.0
colcon-ros==0.5.0
pyasn1==0.4.8
colorama==0.4.4
beniget==0.4.1
colcon-package-information==0.4.0
colcon-powershell==0.4.0
colcon-test-result==0.3.8
distlib==0.3.4
colcon-argcomplete==0.3.3
colcon-common-extensions==0.3.0
colcon-parallel-executor==0.3.0
notify2==0.3
colcon-notification==0.3.0
colcon-devtools==0.3.0
command-not-found==0.3
colcon-cmake==0.2.29
colcon-output==0.2.13
colcon-package-selection==0.2.10
colcon-defaults==0.2.9
colcon-python-setup-py==0.2.9
colcon-metadata==0.2.5
colcon-recursive-crawl==0.2.3
colcon-cd==0.2.1
colcon-installed-package-information==0.2.1
pyasn1-modules==0.2.1
colcon-library-path==0.2.1
unattended-upgrades==0.1
colcon-pkg-config==0.1.0
gyp==0.1
colcon-override-check==0.0.1
mpmath==0.0.0